requirements.txt
streamlit==1.24.0
openai==0.27.8
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pydantic==2.0.3
python-decouple==3.8